refactor:   - AStarPathfinding.Class moved to pathfinding_algorithms   - SimulationCommandHandler.Class added exit command handling   - Position.Class replaced by record   - SimulationMapRenderer.Class added display info about exit comand
